+++ package.json
@@ -26,13 +26,16 @@
   "dependencies": {
     "classnames": "^2.3.1",
     "critters": "^0.0.10",
+    "lodash": "^4.17.21",
     "luxon": "^2.0.1",
+    "mee6-levels-api": "^1.3.0",
     "next": "^11.0.1",
     "react": "^17.0.2",
     "react-dom": "^17.0.2"
   },
   "devDependencies": {
     "@next/bundle-analyzer": "^11.0.1",
+    "@types/lodash": "^4.14.171",
     "@types/luxon": "^1.27.1",
     "@types/node": "^14.17.5",
     "@types/react": "^17.0.14",
