+++ src/pages/index.tsx
@@ -1,6 +1,8 @@
 import * as React from 'react';
 import { GetStaticPropsResult } from 'next';
 
+import { DiscordUser, getTopDiscordUsers } from 'src/api/discord-users';
+import { getGithubStats, GithubStats } from 'src/api/github-stats';
 import {
   AFamiliarExperience,
   Announcement,
@@ -16,9 +18,15 @@ import {
 } from 'src/components';
 import { MinArt } from 'src/components/MinArt';
 
-type Props = React.ComponentPropsWithoutRef<typeof MonthlyGithubAnalytics>;
+type Props = {
+  githubStats: GithubStats;
+  discordUsers: DiscordUser[];
+};
 
 export default function HomePage(props: Props): React.ReactElement {
+  // TODO (HIEU): Add top Discord users
+  console.log(props);
+
   return (
     <>
       <Announcement href="https://twitter.com/MinswapDEX/status/1418221475681558529">
@@ -40,7 +48,7 @@ export default function HomePage(props: Props): React.ReactElement {
 
         <div className="h-20"></div>
 
-        <MonthlyGithubAnalytics {...props} />
+        <MonthlyGithubAnalytics {...props.githubStats} />
 
         <div className="h-20"></div>
 
@@ -66,78 +74,11 @@ export default function HomePage(props: Props): React.ReactElement {
   );
 }
 
-async function fetchRepoStats({ owner, repo, from, to }: { owner: string; repo: string; from: string; to: string }) {
-  const response = await fetch('https://api.github.com/graphql', {
-    method: 'POST',
-    headers: {
-      Authorization: `Bearer ${process.env.GITHUB_TOKEN}`,
-    },
-    body: JSON.stringify({
-      query: `
-        query {
-          repository(owner: "${owner}", name: "${repo}") {
-            object(expression: "main") {
-              ... on Commit {
-                history(since: "${new Date(from).toISOString()}") {
-                  totalCount
-                  nodes {
-                    deletions
-                    additions
-                  }
-                }
-              }
-            }
-          }
-          
-          search(first: 100, query: "is:pr is:merged repo:${owner}/${repo} created:${from}..${to}", type: ISSUE) {
-            issueCount
-          }
-        }
-      `,
-    }),
-  });
-
-  const json = await response.json();
-  const totalCommits = json.data.repository.object.history.totalCount;
-  const totalCodeAddition = json.data.repository.object.history.nodes.reduce(
-    (acc: any, node: any) => acc + node.additions,
-    0,
-  );
-  const totalCodeDeletion = json.data.repository.object.history.nodes.reduce(
-    (acc: any, node: any) => acc + node.deletions,
-    0,
-  );
-  const totalMergedPRs = json.data.search.issueCount;
-
-  return { totalCommits, totalCodeDeletion, totalCodeAddition, totalMergedPRs };
-}
-
 export async function getStaticProps(): Promise<GetStaticPropsResult<Props>> {
-  const { DateTime } = await import('luxon');
-  const now = DateTime.now();
-  const lastMonth = now.minus({ month: 1 });
-  const firstDayLastMonth = lastMonth.startOf('month').toFormat('yyyy-MM-dd');
-  const lastDayLastMonth = lastMonth.endOf('month').toFormat('yyyy-MM-dd');
-
-  const allRes = await Promise.all([
-    fetchRepoStats({ owner: 'longngn', repo: 'minswap-core', from: firstDayLastMonth, to: lastDayLastMonth }),
-    fetchRepoStats({ owner: 'minswap', repo: 'minswap-interface', from: firstDayLastMonth, to: lastDayLastMonth }),
-    fetchRepoStats({ owner: 'minswap', repo: 'minswap-org', from: firstDayLastMonth, to: lastDayLastMonth }),
-  ]);
-
-  const result = allRes.reduce<Props>(
-    (obj, res: Props) => {
-      for (const [key, value] of Object.entries(res)) {
-        obj[key as keyof Props] += value;
-      }
-
-      return obj;
-    },
-    { totalMergedPRs: 0, totalCommits: 0, totalCodeDeletion: 0, totalCodeAddition: 0 },
-  );
+  const [githubStats, discordUsers] = await Promise.all([getGithubStats(), getTopDiscordUsers()]);
 
   return {
-    props: result,
+    props: { githubStats, discordUsers },
     revalidate: 3600 * 24, // Cache for 1 day
   };
 }
